requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asyncer==0.0.2
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
certifi==2023.11.17
chainlit==1.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
fastapi==0.100.1
fastapi-socketio==0.0.10
filetype==1.2.0
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
geocoder==1.38.1
googleapis-common-protos==1.62.0
googlesearch-python==1.2.3
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
html2text==2020.1.16
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.11.0
joblib==1.3.2
Lazify==0.4.0
literalai==0.0.1
llama-index==0.9.26
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.3
openai==1.3.4
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
pandas==2.1.4
protobuf==4.25.1
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyshorteners==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.11.0
pytz==2023.3.post1
ratelim==0.1.6
regex==2023.12.25
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.27.0
syncer==2.0.3
tenacity==8.2.3
tiktoken==0.5.2
tomli==2.0.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.4
uptrace==1.22.0
urllib3==2.1.0
uvicorn==0.23.2
watchfiles==0.20.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
